requirements.txt
asgiref==3.7.2
Django==4.2.5
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1